fix: notification bugs fixed
